backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database Drivers
psycopg2-binary==2.9.9
pymongo==4.6.1

# Environment Variables
python-dotenv==1.0.0

# Data Validation
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0
